remove blocks and txs mixins
