--- starport/ui/src/mixins/blocks/initConnection.js
@@ -1,10 +1,0 @@
-import { mapGetters, mapActions, mapMutations } from 'vuex'
-
-export default {
-  methods: {
-    ...mapActions('cosmos/blocks', [ 'initBlockConnection' ]),
-  },
-  created() {
-    this.initBlockConnection()
-  }
-}
