+++ starport/ui/src/mixins/initCosmos.js
@@ -1,6 +1,5 @@
-import blockConnection from './blocks/initConnection'
 import backendConnection from './backend/initConnection'
 
 export default {
-  mixins: [backendConnection, blockConnection]
+  mixins: [backendConnection]
 }
