--- starport/ui/src/mixins/transactions/init.js
@@ -1,10 +1,0 @@
-import { mapActions } from 'vuex'
-
-export default {
-  methods: {
-    ...mapActions('cosmos/transactions', [ 'initTxsStack' ])
-  },
-  created() {
-    this.initTxsStack()
-  }
-}
