+++ starport/ui/src/modules/BlocksExplorer.vue
@@ -27,7 +27,7 @@
 
 <script>
 import { mapGetters, mapMutations, mapActions } from 'vuex'
-import blockHelpers from '@/mixins/blocks/helpers'
+import { formatter as blockFormatter } from '@/helpers/block'
 
 import _ from 'lodash'
 
@@ -45,7 +45,7 @@ export default {
   },
   data() {
     return {
-      blockFormatter: blockHelpers.blockFormatter(),
+      blockFormatter,
       states: {
         isHidingBlocksWithoutTxs: false,
         isScrolledInTopHalf: true,
